+++ pydatemm/__init__.py
@@ -2,6 +2,6 @@
 
 __author__ = """Thejasvi Beleyur"""
 __email__ = 'thejasvib@gmail.com'
-__version__ = '0.0.1'
+__version__ = '0.0.2'
 
 from pydatemm.source_generation import generate_candidate_sources