+++ HISTORY.rst
@@ -5,4 +5,9 @@ History
 0.0.1 (2022-04-22)
 ------------------
 
-* First release on PyPI.
+* Initial version 
+
+0.0.2 (2023-04-25)
+-------------------
+* Version with command line interface for batch runs
+* Implement environmental variable based libiomp5 library seeking